Make settings an own webview, fix path error in windows
